Check to avoid duplicate work

+++ run_icpe21_experiments.py
@@ -25,6 +25,9 @@ for folder in next(os.walk(trace_dir))[1]:
 
     for tu, tsp, tpp, dvfs_enabled in itertools.product(target_utilizations, task_selection_policies,
                                                         task_placement_policies, dvfs_enabled):
+        output_dir = os.path.join(output_location, f"{folder}_tu_{tu}_tsp_{tsp}_tpp_{tpp}_dvfs_{dvfs_enabled}")
+        if os.path.exists(output_dir):
+            continue
         command = "java -cp target/wta-sim-0.1.jar science.atlarge.wta.simulator.WTASim -f wta"
         command += " -c " + " ".join([str(x) for x in machine_resources])
         command += " -t " + " ".join([str(x) for x in machine_tdps])
@@ -32,7 +35,7 @@ for folder in next(os.walk(trace_dir))[1]:
         command += " -mf " + " ".join([str(x) for x in machine_fractions])
         command += " -e " + " ".join([str(x) for x in [dvfs_enabled] * len(machine_resources)])
         command += " -i " + os.path.join(trace_dir, folder)
-        command += " -o " + os.path.join(output_location, f"{folder}_tu_{tu}_tsp_{tsp}_tpp_{tpp}_dvfs_{dvfs_enabled}")
+        command += " -o " + output_dir
         command += " --target-utilization " + str(tu)
         command += " --task-order-policy " + tsp
         command += " --task-placement-policy " + tpp
